AVRO-346.  Add function to validate a datum against a schema
git-svn-id: https://svn.apache.org/repos/asf/hadoop/avro/trunk@900279 13f79535-47bb-0310-9956-ffa450edef68

+++ lang/c/src/io.c
@@ -123,7 +123,7 @@ avro_writer_file (FILE * fp)
 }
 
 avro_reader_t
-avro_reader_memory (const char *buf, size_t len)
+avro_reader_memory (const char *buf, int64_t len)
 {
   struct avro_memory_reader_t *mem_reader =
     malloc (sizeof (struct avro_memory_reader_t));
@@ -139,7 +139,7 @@ avro_reader_memory (const char *buf, size_t len)
 }
 
 avro_writer_t
-avro_writer_memory (const char *buf, size_t len)
+avro_writer_memory (const char *buf, int64_t len)
 {
   struct avro_memory_writer_t *mem_writer =
     malloc (sizeof (struct avro_memory_writer_t));
